Refactor login route
